Use float dtype for passenger count
In order to handle NaN entries in the passenger count column, use a float dtype. Note, Pandas Int64 is not appropriate here as it makes xgboost
think the column should be treated as categoriacal

+++ tip_prediction.py
@@ -59,7 +59,7 @@ def load_raw_data():
         dtype={
             "payment_type": "UInt8",
             "VendorID": "UInt8",
-            "passenger_count": "Int64",
+            "passenger_count": "float",
             "RatecodeID": "UInt8",
             "store_and_fwd_flag": "category",
             "PULocationID": "UInt16",
